backend/requirements.txt
Django>=5.0,<5.2
djangorestframework>=3.15,<3.16
django-cors-headers>=4.4,<4.6
psycopg2-binary>=2.9,<2.10
dj-database-url>=2.2,<2.3

pytest>=8.2,<9.0
pytest-django>=4.9,<5.0
pytest-cov>=5.0,<6.0
